WebUI/requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
cffi==0.8.6
cryptography==0.7.1
enum34==1.0.4
httplib2==0.9
identity-toolkit-python-client>=0.1.6
itsdangerous==0.24
oauth2client==1.4.5
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
wsgiref==0.1.2
web.py==0.37
MySQL-python==1.2.5